+++ Content.Server/Chemistry/EntitySystems/ChemistrySystemHypospray.cs
@@ -1,9 +1,20 @@
 using System.Linq;
+using System.Diagnostics.CodeAnalysis;
 using Content.Server.Chemistry.Components;
+using Content.Server.Chemistry.Components.SolutionManager;
+using Content.Server.Interaction.Components;
+using Content.Server.Weapons.Melee;
+using Content.Shared.Chemistry.Reagent;
+using Content.Shared.Database;
+using Content.Shared.FixedPoint;
+using Content.Shared.IdentityManagement;
 using Content.Shared.Interaction;
 using Content.Shared.Interaction.Events;
+using Content.Shared.MobState.Components;
+using Content.Shared.Popups;
 using Content.Shared.Weapons.Melee;
 using Content.Shared.Weapons.Melee.Events;
+using Robust.Shared.Player;
 
 namespace Content.Server.Chemistry.EntitySystems
 {
@@ -21,7 +32,7 @@ namespace Content.Server.Chemistry.EntitySystems
         {
             if (args.Handled) return;
 
-            component.TryDoInject(args.User, args.User);
+            TryDoInject(component, args.User, args.User);
             args.Handled = true;
         }
 
@@ -30,7 +41,7 @@ namespace Content.Server.Chemistry.EntitySystems
             Dirty(component);
         }
 
-        public void OnAfterInteract(EntityUid uid, HyposprayComponent comp, AfterInteractEvent args)
+        public void OnAfterInteract(EntityUid uid, HyposprayComponent component, AfterInteractEvent args)
         {
             if (!args.CanReach)
                 return;
@@ -38,15 +49,101 @@ namespace Content.Server.Chemistry.EntitySystems
             var target = args.Target;
             var user = args.User;
 
-            comp.TryDoInject(target, user);
+            TryDoInject(component, target, user);
         }
 
-        public void OnAttack(EntityUid uid, HyposprayComponent comp, MeleeHitEvent args)
+        public void OnAttack(EntityUid uid, HyposprayComponent component, MeleeHitEvent args)
         {
             if (!args.HitEntities.Any())
                 return;
 
-            comp.TryDoInject(args.HitEntities.First(), args.User);
+            TryDoInject(component, args.HitEntities.First(), args.User);
+        }
+
+        public bool TryDoInject(HyposprayComponent component, EntityUid? target, EntityUid user)
+        {
+            if (!EligibleEntity(target, _entMan))
+                return false;
+
+            string? msgFormat = null;
+
+            if (target == user)
+            {
+                msgFormat = "hypospray-component-inject-self-message";
+            }
+            else if (EligibleEntity(user, _entMan) && _interaction.TryRollClumsy(user, component.ClumsyFailChance))
+            {
+                msgFormat = "hypospray-component-inject-self-clumsy-message";
+                target = user;
+            }
+
+            var solutionsSys = EntitySystem.Get<SolutionContainerSystem>();
+            solutionsSys.TryGetSolution(component.Owner, component.SolutionName, out var hypoSpraySolution);
+
+            if (hypoSpraySolution == null || hypoSpraySolution.CurrentVolume == 0)
+            {
+                user.PopupMessageCursor(Loc.GetString("hypospray-component-empty-message"));
+                return true;
+            }
+
+            if (!solutionsSys.TryGetInjectableSolution(target.Value, out var targetSolution))
+            {
+                user.PopupMessage(user,
+                                  Loc.GetString("hypospray-cant-inject", ("target", Identity.Entity(target.Value, _entMan))));
+                return false;
+            }
+
+            user.PopupMessage(Loc.GetString(msgFormat ?? "hypospray-component-inject-other-message",
+                                            ("other", target)));
+            if (target != user)
+            {
+                target.Value.PopupMessage(Loc.GetString("hypospray-component-feel-prick-message"));
+                var meleeSys = EntitySystem.Get<MeleeWeaponSystem>();
+                var angle = Angle.FromWorldVec(_entMan.GetComponent<TransformComponent>(target.Value).WorldPosition - _entMan.GetComponent<TransformComponent>(user).WorldPosition);
+                // TODO: This should just be using melee attacks...
+                // meleeSys.SendLunge(angle, user);
+            }
+
+            _audio.PlayPvs(component.InjectSound, user);
+
+            // Get transfer amount. May be smaller than _transferAmount if not enough room
+            var realTransferAmount = FixedPoint2.Min(component.TransferAmount, targetSolution.AvailableVolume);
+
+            if (realTransferAmount <= 0)
+            {
+                user.PopupMessage(user,
+                                  Loc.GetString("hypospray-component-transfer-already-full-message",
+                                                ("owner", target)));
+                return true;
+            }
+
+            // Move units from attackSolution to targetSolution
+            var removedSolution =
+            EntitySystem.Get<SolutionContainerSystem>()
+            .SplitSolution(component.Owner, hypoSpraySolution, realTransferAmount);
+
+            if (!targetSolution.CanAddSolution(removedSolution))
+            {
+                return true;
+            }
+
+            removedSolution.DoEntityReaction(target.Value, ReactionMethod.Injection);
+
+            EntitySystem.Get<SolutionContainerSystem>().TryAddSolution(target.Value, targetSolution, removedSolution);
+
+            _adminLogger.Add(LogType.ForceFeed,
+                             $"{_entMan.ToPrettyString(user):user} injected {_entMan.ToPrettyString(target.Value):target} with a solution {SolutionContainerSystem.ToPrettyString(removedSolution):removedSolution} using a {_entMan.ToPrettyString(component.Owner):using}");
+
+            return true;
+        }
+
+        static bool EligibleEntity([NotNullWhen(true)] EntityUid? entity, IEntityManager entMan)
+        {
+            // TODO: Does checking for BodyComponent make sense as a "can be hypospray'd" tag?
+            // In SS13 the hypospray ONLY works on mobs, NOT beakers or anything else.
+
+            return entMan.HasComponent<SolutionContainerManagerComponent>(entity)
+            && entMan.HasComponent<MobStateComponent>(entity);
         }
     }
 }
